修改bug，修改字段类型
1、bug，在批次的状态status 设置为2时，积分查询、名单公示页面报错，需要修正；
2、分数类型修改为double类型，因为分数线，有可能包含小数；

+++ src/main/java/com/wutuobang/score/model/BatchConfModel.java
@@ -73,7 +73,7 @@ public class BatchConfModel implements Serializable {
 	/*
 	2019年1月4日
 	 */
-	private Integer scoreValue;//"指标值"
+	private Double scoreValue;//"指标值"
 
 
 	//columns END
@@ -254,11 +254,11 @@ public class BatchConfModel implements Serializable {
 		return openLoginTime;
 	}
 
-	public Integer getScoreValue() {
+	public Double getScoreValue() {
 		return scoreValue;
 	}
 
-	public void setScoreValue(Integer scoreValue) {
+	public void setScoreValue(Double scoreValue) {
 		this.scoreValue = scoreValue;
 	}
